render: split render.h into wlr_renderer.h and wlr_texture.h
