+++ examples/simple.c
@@ -1,11 +1,11 @@
 #define _POSIX_C_SOURCE 199309L
-#include <string.h>
+#include <GLES2/gl2.h>
+#include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include <time.h>
-#include <inttypes.h>
 #include <wayland-server.h>
-#include <GLES2/gl2.h>
 #include <wlr/backend.h>
 #include <wlr/backend/session.h>
 #include <wlr/types/wlr_output.h>
