+++ render/gles2/texture.c
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <wayland-server-protocol.h>
 #include <wayland-util.h>
-#include <wlr/render.h>
+#include <wlr/render/wlr_texture.h>
 #include <wlr/render/egl.h>
 #include <wlr/render/interface.h>
 #include <wlr/types/wlr_matrix.h>
