+++ types/wlr_screenshooter.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <wayland-server.h>
 #include <wlr/backend.h>
-#include <wlr/render.h>
+#include <wlr/render/wlr_renderer.h>
 #include <wlr/types/wlr_output.h>
 #include <wlr/types/wlr_screenshooter.h>
 #include <wlr/util/log.h>
